image-augmentation-aws/lambda-functions/image-processor/requirements.txt
Pillow==10.0.1
boto3==1.28.57
botocore==1.31.57
